:recycle: Move study mutations to new file
